pyproject.toml
[project]
name = "imagegen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.1.0",
    "datasets>=3.6",
    "deeplake>=4.2.4",
    "diffusers>=0.33.1",
    "flake8>=7.2.0",
    "h5py>=3.13.0",
    "hf-xet>=1.1.2",
    "huggingface-hub>=0.31.2",
    "img2dataset>=1.45.0",
    "ipdb>=0.13.13",
    "ipython>=9.2.0",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.3",
    "pandas>=2.2.3",
    "pillow>=11.2.1",
    "pytest>=8.3.5",
    "pytorch-model-summary>=0.1.2",
    "scikit-learn>=1.6.1",
    "scipy>=1.15.3",
    "tensorboard>=2.19.0",
    "torch>=2.7.0",
    "torchvision>=0.22.0",
    "transformers>=4.51.3",
]
[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
